backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.3.1
python-dotenv==1.1.1
regex==2024.11.6
scikit-learn==1.7.0
scipy==1.16.0
threadpoolctl==3.6.0
tqdm==4.67.1
Werkzeug==3.1.3